requirements.txt
Pillow==10.0.1
pyinstaller==5.9.0